Refactor test queries file
