+++ photo/tests/test_queries/graphql_queries.py
